# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: EVC Mesh API
  description: |
    REST API for EVC Mesh -- a multi-agent task management platform.
    Provides workspace, project, task, comment, artifact, agent, and event-bus
    endpoints for both human users (JWT) and AI agents (API key).
  version: "0.1.0"
  contact:
    name: EVC Mesh Team
servers:
  - url: http://localhost:8005
    description: Local development
security:
  - BearerAuth: []
  - AgentKeyAuth: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via /auth/login or /auth/refresh.
    AgentKeyAuth:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key in the format agk_{workspace_slug}_{random}.

  # ------------------------------------------------------------------
  # Reusable schemas
  # ------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties:
            type: string

    # ---------- Pagination wrapper ----------
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        agent_type:
          type: string
          enum: [claude_code, openclaw, cline, aider, custom]
        api_key_prefix:
          type: string
        capabilities:
          type: object
        status:
          type: string
          enum: [online, offline, busy, error]
        last_heartbeat:
          type: string
          format: date-time
          nullable: true
        current_task_id:
          type: string
          format: uuid
          nullable: true
        total_tasks_completed:
          type: integer
        total_errors:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/agents/me:
    get:
      tags: [Agents]
      summary: Get the current agent's own profile
      operationId: getAgentMe
      security:
        - AgentKeyAuth: []
      responses:
        "200":
          description: Current agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          description: Unauthorized (agent API key required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Agents]
      summary: Update the current agent's own profile (self-service)
      operationId: updateAgentMe
      security:
        - AgentKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile_description:
                  type: string
                callback_url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Updated agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...